Added authorization to user routes

+++ routes/product.routes.js
@@ -1,21 +1,18 @@
-// Importamos express:
 const express = require("express");
-// Importamos el modelo que nos sirve tanto para importar datos como para leerlos:
-const { Product } = require("../models/Product.js");
-// Importamos la función que nos sirve para resetear los product:
-const { resetProducts } = require("../utils/resetProducts.js");
-
-// Importamos la función que nos sirve para resetear los users:
-const { resetUsers } = require("../utils/resetUsers.js");
-
-// Importamos la función que nos sirve para resetear las relaciones entre las coleciones
-// Router propio de product suministrado por express.Router:
-const router = express.Router();
-// Importamos Multer para subir photos.
-const multer = require("multer");
-const upload = multer({ dest: "public" });
-// Import filesystem "fs"
 const fs = require("fs");
+const multer = require("multer");
+
+const { Product } = require("../models/Product.js");
+
+const { resetProducts } = require("../utils/resetProducts.js");
+const { resetUsers } = require("../utils/resetUsers.js");
+const { resetChats } = require("../utils/resetChats.js");
+
+const { isAuth } = require("../middlewares/auth.middleware");
+
+const router = express.Router();
+
+const upload = multer({ dest: "public" });
 
 // --------------------------------------------------------------------------------------------
 // --------------------------------- ENDPOINTS DE /product ---------------------------------------
@@ -113,9 +110,14 @@ router.get("/name/:name", async (req, res, next) => {
 
 //  Endpoint para añadir elementos (CRUD: CREATE):
 
-router.post("/", async (req, res, next) => {
+router.post("/", isAuth, async (req, res, next) => {
   // Si funciona la escritura...
   try {
+    const id = req.params.id; //  Recogemos el id de los parametros de la ruta.
+
+    if (req.user.id !== id && req.user.email !== "admin@gmail.com") {
+      return res.status(401).json({ error: "No tienes autorización para realizar esta operación" });
+    }
     const product = new Product(req.body); //     Un nuevo product es un nuevo modelo de la BBDD que tiene un Scheme que valida la estructura de esos datos que recoge del body de la petición.
     const createdProduct = await product.save(); // Esperamos a que guarde el nuevo product creado en caso de que vaya bien. Con el metodo .save().
     return res.status(201).json(createdProduct); // Devolvemos un código 201 que significa que algo se ha creado y el product creado en modo json.
@@ -133,8 +135,13 @@ router.post("/", async (req, res, next) => {
 //  Endpoint para asociar una imágen a una user:
 //  Hacemos uso del middleware que nos facilita multer para guardar la imágen en la carpeta de estáticos public.
 
-router.post("/image-upload/:id", upload.single("image"), async (req, res, next) => {
+router.post("/image-upload/:id", isAuth, upload.single("image"), async (req, res, next) => {
   try {
+    const id = req.params.id; //  Recogemos el id de los parametros de la ruta.
+
+    if (req.user.id !== id && req.user.email !== "admin@gmail.com") {
+      return res.status(401).json({ error: "No tienes autorización para realizar esta operación" });
+    }
     const productId = req.params.id;
     const product = await Product.findById(productId);
 
@@ -175,8 +182,9 @@ router.delete("/reset", async (req, res, next) => {
 
     // Si all es true resetearemos todos los datos de nuestras coleciones y las relaciones entre estas.
     if (all) {
-      await resetProducts();
       await resetUsers();
+      await resetProducts();
+      await resetChats();
       res.send("Datos reseteados y Relaciones reestablecidas");
     } else {
       await resetProducts();
@@ -192,10 +200,14 @@ router.delete("/reset", async (req, res, next) => {
 
 //  Endpoint para eliminar product identificado por id (CRUD: DELETE):
 
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", isAuth, async (req, res, next) => {
   // Si funciona el borrado...
   try {
     const id = req.params.id; //  Recogemos el id de los parametros de la ruta.
+
+    if (req.user.id !== id && req.user.email !== "admin@gmail.com") {
+      return res.status(401).json({ error: "No tienes autorización para realizar esta operación" });
+    }
     const productDeleted = await Product.findByIdAndDelete(id); // Esperamos a que nos devuelve la info del product eliminado que busca y elimina con el metodo findByIdAndDelete(id del product a eliminar).
     if (productDeleted) {
       res.json(productDeleted); //  Devolvemos el product eliminado en caso de que exista con ese id.
@@ -218,10 +230,15 @@ fetch("http://localhost:3000/product/id del product a borrar",{"method":"DELETE"
 
 //  Endpoint para actualizar un elemento identificado por id (CRUD: UPDATE):
 
-router.put("/:id", async (req, res, next) => {
+router.put("/:id", isAuth, async (req, res, next) => {
   // Si funciona la actualización...
   try {
     const id = req.params.id; //  Recogemos el id de los parametros de la ruta.
+
+    if (req.user.id !== id && req.user.email !== "admin@gmail.com") {
+      return res.status(401).json({ error: "No tienes autorización para realizar esta operación" });
+    }
+
     const productUpdated = await Product.findByIdAndUpdate(id, req.body, { new: true, runValidators: true }); // Esperamos que devuelva la info del product actualizado al que tambien hemos pasado un objeto con los campos q tiene que acualizar en la req del body de la petición. {new: true} Le dice que nos mande el product actualizado no el antiguo. Lo busca y elimina con el metodo findByIdAndDelete(id del product a eliminar).
     if (productUpdated) {
       res.json(productUpdated); //  Devolvemos el product actualizado en caso de que exista con ese id.
